Checkpoint

+++ src/ldak/decomp.h
@@ -3,8 +3,8 @@
 double cg_solve(double *mat, int length, double *mat2, double *mat3, int ncol,
                 double tol);
 
-double eigen_invert(double *mat, int length, double *mat2, int ncol,
-                    double *mat3, int type);
+int eigen_invert(double *mat, int length, double *mat2, int ncol, double *mat3,
+                 int type);
 
 void eigen_strip(double *mat, int start, int end, int total, double strip);
